First bash at a three-screen UI
This 'sort of' works. There's a nasty bug with the handlers though. What appears
to be happening is that multiple instances of something are causing the handler to trigger
multiple times on each page switch. This materialises lines being skipped.

Aborting on  this idea (Init and Close for each page) for now.