Support page scrolling in TicketShow and TicketList pages
This is pretty noddy, but at least it allows for scrolling now.

I can't find a way of getting the 'adjusted line count' for a ui.List in 'wrap' mode,
so I've added a 'correction' factor to the calculation of the last line.

This is not going to work in all cases, so I really need to add some code that works out just
how many lines have been wrapped :(