Embrace structs, start making code more consistent
This rearranges the QueryPage code into a struct with methods, with
a view to making all pages fit this form.

It should then allow for simplifying the main logic significantly, and
definitely will remove the need for stupid per-page globla variables.

No functional change is made here (bar I think a bugfix with previousLine)