Add LabelListPage -- label-centred view on a Query
This adds LabelListPage, which takes the results of a query and lists the labels that are
used by tickets within it.

This uses the direct output of jira.FindIssues(), which is pretty gnarly coming from the JSON
returned by the API. However, this does mean that we do not depend on the "labels()" section
in TicketsListPage. Win!

The intention is to be able to select a label and see the subquery including just tickets of
that label. However, this involves a bit of a rearrange, so I'm leaving out for now.