Allow opts to be passed to Create(); Support subqueries on labels
I'm not convinced that this is the right way to go, but this commit
does add support for displaying a subset of results on the TicketList page based on
coming from the LabelList page - so you can show tickets matching the query + label.

This is extensible, so is good in some ways. However, it feels messy. The only alternative
I can think of is to have a separate 'LabelLimitedTicketListPage' or something. Meh.