Add edit support, with caveat - need to exit after
'E' edits the current ticket on both List and Show pages (basically where we
have an obvious ticketId available)

This fixes #2, but opens #8 - CmdEdit() requires that we close down
termui, as it needs control of the keyboard (fine).

However, I've not yet found a way of implementing this so that we can
subsequently re-enter the termui control. Meh. Hence #8.

Same problemw will happen with all editor based operations, #8 covers them.