Add commenting support, badly
This fixes #9, though still suffering from #8.

I'm not particularly happy here though, the edit interface provided lacks context: it simply
is a raw YAML 'body: |' prompt in $EDITOR. Not very friendly.

Unfortunately, the data is just not available to the 'comment' template to put a reference
to the ticket id and summary into place, so I think this requires a PR to go-jira :(