Add ticket trail handling in TicketShowPage
When drilling into tickets mentioned in the Show page, it's annoying to GoBack
to the List page. This implements a breadcrumb trail that means we can GoBack to the
previous ticket viewed in the trail.

This is relying on the immutable TicketShowPage change in prior commits - we don't need to
remove from the trail, because we're just moving back to the previous page (which automatically
has the shorter trail).