Add navigation to go to bottom and top of page
Found I was needing this - it's common to want to jump to the end of a ticket to
get the latest context. Returning to the top is handy for getting to the description and
metadata.

Also handy as a general navigation to get to the bottom of the query/list pages.