Use our own text wrapping in TicketShowPage
As described in #5, the termui List wrapping doesn't appear to give a
method to access the adjusted 'interior line count' after wrapping text.

This implements our own version, which at present is still pretty dumb:

* Does not wrap single line field values, like 'summary:' Mostly because I
  think this will help - truncating is better.
* Hard wraps on the maxWidth - no hyphening or word boundary matching.

I've given a default maxWidth of 78, mostly because it aides readability. I
tried 130, and it made it harder to join up wrapped lines in my head. YMMV. It
should probably be made configurable later.

This is 'good enough'. It isn't pretty, but is better than what we had. Paging
now works intuitively. Yey.