Allow for [](fg-red) style colouring of lines
We previously wrapped every line with [](), which prevented templates
from using the [highlight](colour) termui syntax themselves.

This is a little clunky, in that it clashes slightly with the selected line
highlight (which takes precedence, and as a result shows the inner markup), but
this is less annoying than non-coloured text.

I think it's also introduced a bug in displaying JIRA markdown links, but meh.