Fixes #11 - Support drilling into epics
We dont get the epic sub-issues via the API call to the Epic issue. Instead
provide a workaround to view them via a query (unresolved only btw)

This adds secret sauce to the Select() call for TicketShow that matches on a
'^epic_links' field in the template, which then produces a query into the TicketList
page.

Bit cludgy, but it works.