Add 'move by paragraph' keys -- { and }
I love paragraph movement in Vim, so I'm implementing it in this.

Default to 5 line 'rapid' movement where paragraphs don't make sense.

However, for TicketShow, we have paragraphs -- move to the next blank line.

And for QueryPage, we have 'spacer lines' (where there is no JQL in the Query). Move
to the next line without any JQL. This works great!