Implement a 'display cursor' on blank lines
A UX problem with move-by-paragraph is that the cursor disappears when moving to
blank lines. This wasn't so big a deal when we were moving line by line, but it affects
every paragraph move - so you rapidly lose where you are.

Work around this by displaying a single space instead of a blank line. You then get a nice
single char cursor. Woo!