Support refreshing the page.
This is sort of needed, tickets are going to get updated whilst we're
looking at them, and new tickets will be added. We should be able to get new
results/comments/updates that we know of without having to back up/reenter.

Seems to work pretty well. Interestingly this results in a call to BaseListPage.Create()
and then (thankfully) calls the actual page Create() when the data is loaded. Probably
something to do with Update().

For now, changing the title of the Base window to 'Updating...' makes this a feature ;)