Allow custom list templates on per-query basis
Some queries are more easy to process with a custom list template, eg voter counts.

This commit adds support for per-query templates via the jira-ui-config.yml, eg:

    queries:
      - name: project votes
        jql:  project = TEST AND resolution = Unresolved ORDER BY votes,created
        template: colour_list

This is surprisingly powerful.

The template must be created in the .jira.d/templates/ path as normal. Examples are
provided in the `examples/templates` folder.

Colourising the templates (highly useful) caused an issue with the TicketId grabbing, so
I've loosened that to use findTicketInString() - shouldn't be an issue generally.