Fix resize wrapping in TicketShow; cache API body
Previously we were caching the post-template API body, into
p.CachedResult. This was also word-wrapped.

It's better to cache the API body, and reprocess as needed.

Note that this also fixes up resetting the position of the cursor,
so it's preserved across resize operations.

We need to override Refresh() now, to ensure we clear down apiBody.