Add sort ordering to TicketList via SortOrderPage!
Fixes #7

Implement sort ordering, via a SortOrderPage 'popup'. <3 termui.

This works in a very similar way to the QueryPage, in fact is almost a
copy-pasta of it.

Basically, sorts in JIRA are just a bit of JQL appended to the query. So
it's not surprising that SortOrderPage =~ QueryPage

Extend TicketList with the notion of an ActiveSort, which is then appended
to the ActiveQuery.JQL.

Add a bit of logic to strip any existing ORDER BY clause, because we don't
want to double up (it breaks the JQL).

A 'default' sort basically just doesn't add the ORDER BY to the JQL. It does
append a blank space though, as we load the meaning of "" for no-op spacer lines
in the UI (eg for paragraph movement). Dirty hack, but meh, simple to implement.