Implement a 'working' command bar in QueryPage
This gets a command bar that QueryPage can actually react to, in
fact ':q<enter>' is implemented.

Refactored the ui_control.go handlers to be clearer. Still more work
to be done there, but the way it was implemented made the editbox key
handling hard. Now it's straightforward and easy to read/add-to.