Implement assigee commands: :assign, :unassign, :take
These (bar unassign, todo) map to the respective go-jira commands.

unassign updates with 'assignee: { name: -1 }' -- which set the assignee to the
default. Technically this could be a user, but the point is the same - unassign from
the current user, throw back into the pool. In most cases, this will make it 'Unassigned'