Add Searching in all main pages
This connects the dots with the Search function to do a regexp match
on the relevant portion of cachedResults.

    /{regex} searches down
    ?{regex} searches up

'n' selects the next result (up or down, depending on original search)

This highlighted some bugs with travelling to specific lines in the  output,
which are fixed in this commit (hence the FixDisplayedLine() refactor)

QueryPage continues to be a snowflake, because it has its own cachedResults slice.