Remove externally edited comment support
This was grim, and buggy. 'C' is a bad choice for a command (i think it somehow
conflicts with Ctrl-{key} commands?), but largely down to the external edit operation
causing a quit - and that we now can comment via :comment command.

I'll reimplement soon with its own page, for multi-line comments. For now though, bye!