Include unlabelled tickets in LabelListPage.
it's pretty lame to not be able to see unlabelled tickets when looking
at label view. This commit categorizes them into "NOT LABELLED", which is then
treated specially for the drill down (since we need to append 'labels IS EMPTY'
to the query)

This is a bit hacky, but solves the problem quick.