Add :view {ticket} and :query {JQL} commands; fix display bug.
This adds :view and :query commands, to view adhoc tickets and JQL as needed.

   :view {ticket-id}  -- opens ticket up in TicketShow page
   :query {JQL}       -- opens up TicketList page with custom query

Also adds :help

It highlighted a bug in the CommandBox.Submit() func, as in these cases the
ExecuteCommand call is updating CurrentPage, so we need to refresh our idea of it
when doing the type check.