Implement history for searches and commands
This adds independent history for 'searches' (? and /) and
'commands' (:)

It needs to introduce a commandType into CommandBox, since '/' and '?'
are equivalent from a history perspective (you want to search for the same
thing up and down)

This also fixes an issue where the cursor would be placed at the start of the
command - we now put it at the end. We still need to implement cursor key movement
in this box, but at least this makes it useable.

It also fixes the case where you 'backspace over' the command itself. This now exits
commandMode -- which mirrors vim's behaviour.

There are bugs in this though, but largely it's worky.