Fixes #8 - Allow breaking out of the UI to run external commands (eg vim)
This isn't perfect, but it does work acceptably. By turning off event handlers
(using empty func() as the handler for an event) and by readding echo and opost to the
tty, we get a functional terminal.

(NB: the stty isn't strictly necessary for vim, but does cause a problem if you enter vim
and then get to a shell from inside it. So screw it, make it work for interactive shell too via
:shell command.)